add helper functions js file
