+++ src/components/Student.jsx
@@ -1,24 +1,7 @@
 import React from "react";
 import "./Student.css";
 import { Link } from "react-router-dom";
-
-export function formatBirthDate(inputDate) {
-  const [month, day, year] = inputDate.split('/').map(Number);
-  const monthNames = [
-      "January", "February", "March", "April", "May", "June",
-      "July", "August", "September", "October", "November", "December"
-  ];
-  const formattedDate= `${monthNames[month - 1]} ${day}, ${year}`;
-  return formattedDate;
-}
-
-export function onTrackChecker({codewars:{current:{total}}, certifications:{resume,linkedin,github,mockInterview}}) {
-  if(total >850 && resume && linkedin && mockInterview && github){
-    return true
-  }else{
-    return false
-  }
-}
+import { formatBirthDate, onTrackChecker } from "../helpers/helpers";
 const Student = ({ student }) => {
     const fullName = `${student.names.preferredName} ${student.names.middleName} ${student.names.surname}`
     const birthday = formatBirthDate(student.dob)
